requirements.txt
grpcio==1.73.0
grpcio-tools==1.73.0
pyvisa==1.15.0
